Add exception handling + refactor
