+++ app/clients/aws/client.py
@@ -1,11 +1,9 @@
-import json
 import logging
 import os
 from typing import Any
 
 import boto3
 import botocore.client
-from pydantic import BaseModel
 
 from app.model.aws_config import AWSConfig
 
@@ -17,11 +15,6 @@ _AWS_SECRET_ACCESS_KEY = os.getenv("AWS_SECRET_ACCESS_KEY", "")
 _SIGNATURE_VERSION = "s3v4"
 
 AWSClient = Any
-
-
-class S3UploadConfig(BaseModel):
-    bucket_name: str
-    object_name: str
 
 
 def _get_client_from_config(config: AWSConfig) -> AWSClient:
@@ -45,32 +38,3 @@ def get_s3_client() -> AWSClient:
         region_name=_AWS_REGION,
     )
     return _get_client_from_config(config)
-
-
-def upload_json_to_s3(config: S3UploadConfig, json_data: dict) -> None:
-    """
-    Upload a JSON file to S3
-
-    :param S3UploadConfig config: The configuration required for the upload.
-    :param dict json_data: The json data to be uploaded to S3.
-    """
-    s3_client = get_s3_client()
-    # try:
-    s3_client.put_object(
-        Bucket=config.bucket_name,
-        Key=config.object_name,
-        Body=json.dumps(json_data),
-        ContentType="application/json",
-    )
-    #     logger.info(
-    #         "🎉 Successfully uploaded JSON to S3: %s/%s",
-    #         config.bucket_name,
-    #         config.object_name,
-    #     )
-    # except Exception as e:
-    #     logger.error("💥 Failed to upload JSON to S3: %s", e)
-    #     raise
-
-
-# Example usage:
-# upload_json_to_s3(S3UploadConfig(bucket_name="my-bucket", object_name="data.json"), {"key": "value"})
